+++ packages/embark-utils/index.d.ts
@@ -5,10 +5,13 @@ declare module "embark-utils" {
     prepareForCompilation(isCoverage?: boolean): any;
   }
 
+  function anchoredValue(anchor: string|null, value: string): string;
+  function anchoredPath(anchor: string|null, ...args: string[]): string;
   function compact(array: any): any;
   function checkIsAvailable(url: string, callback: any): void;
   function dockerHostSwap(host: string): string;
   function dappPath(...names: string[]): string;
+  function diagramPath(...names: string[]): string;
   function escapeHtml(message: any): string;
   function embarkPath(...names: string[]): string;
   function exit(code?: any): void;
@@ -17,5 +20,6 @@ declare module "embark-utils" {
   function fuzzySearch(text: string, list: any, filter: any): any;
   function getExternalContractUrl(file: string, provideUrl: string): string;
   function recursiveMerge(target: any, source: any): any;
+  function pkgPath(...names: string[]): string;
   function removePureView(dir: string): void;
 }
