+++ packages/embark/src/lib/core/fs.js
@@ -1,15 +1,11 @@
 /* global module process require */
 
-const {DAPP_PATH,
-       DIAGRAM_PATH,
-       EMBARK_PATH,
-       PKG_PATH,
-       anchoredValue} = require('./env');
 const fs = require('fs-extra');
-const os = require('os');
 const parseJson = require('parse-json');
 const path = require('path');
-import {joinPath, sha512} from 'embark-utils';
+import { joinPath } from 'embark-utils';
+import './env';
+
 require('colors');
 
 function mkdirpSync(...args) { return fs.mkdirpSync(...args); }
@@ -73,37 +69,7 @@ function access(...args) { return fs.access(...args); }
 
 function removeSync(...args) { return fs.removeSync(...args); }
 
-function anchoredPath(anchor, ...args) {
-  return joinPath(
-    anchoredValue(anchor),
-    ...args.map(path => path.replace(dappPath(), ''))
-  );
-}
-
-function embarkPath(...args) { return anchoredPath(EMBARK_PATH, ...args); }
-
-function dappPath(...args) { return anchoredPath(DAPP_PATH, ...args); }
-
-function diagramPath(...args) { return anchoredPath(DIAGRAM_PATH, ...args); }
-
-function ipcPath(basename, usePipePathOnWindows = false) {
-  if (!(basename && typeof basename === 'string')) {
-    throw new TypeError('first argument must be a non-empty string');
-  }
-  if (process.platform === 'win32' && usePipePathOnWindows) {
-    return `\\\\.\\pipe\\${basename}`;
-  }
-  return joinPath(
-    tmpDir(`embark-${sha512(dappPath()).slice(0, 8)}`),
-    basename
-  );
-}
-
-function pkgPath(...args) { return anchoredPath(PKG_PATH, ...args); }
-
 function createWriteStream(...args) { return fs.createWriteStream(...args); }
-
-function tmpDir(...args) { return joinPath(os.tmpdir(), ...args); }
 
 function copyPreserve(sourceFilePath, targetFilePath) {
   const implementation = (sourceFilePath, targetFilePath) => {
@@ -135,20 +101,15 @@ module.exports = {
   copyPreserve,
   copySync,
   createWriteStream,
-  dappPath,
-  diagramPath,
-  embarkPath,
   existsSync,
   ensureFileSync,
   ensureDirSync,
-  ipcPath,
   mkdirp,
   mkdirpSync,
   move,
   moveSync,
   outputFileSync,
   outputJSONSync,
-  pkgPath,
   readFile,
   readFileSync,
   readJSONSync,
@@ -159,7 +120,6 @@ module.exports = {
   stat,
   statSync,
   symlink,
-  tmpDir,
   writeFile,
   writeFileSync,
   writeJSONSync,
